Move subprocess classes together into one utility boundary
