+++ salt/utils/subprocess.py
@@ -4,12 +4,48 @@ For running command line executables with a timeout
 '''
 from __future__ import absolute_import, print_function, unicode_literals
 
+import os
+import sys
 import shlex
+import logging
 import subprocess
 import threading
 import salt.exceptions
 import salt.utils.data
 from salt.ext import six
+
+
+log = logging.getLogger(__name__)
+
+
+class FdPopen(subprocess.Popen):
+    '''
+    Python2 closing file descriptors in a blind way by a system's given range.
+    Therefore close_fds can significantly slow down the system in various circumstances.
+    '''
+
+    if sys.platform.startswith('linux') and six.PY2:
+        def _close_fds(self, but):
+            '''
+            Close only file descriptors that are needed.
+            The issue is not new: https://bugs.python.org/issue1663329
+
+            :param but:
+            :return:
+            '''
+            maxfds = len(os.listdir('/proc/{}/fd'.format(os.getpid())))
+            log.trace('Closing {} file descriptors'.format(maxfds))
+            if hasattr(os, 'closerange'):
+                os.closerange(3, but)
+                os.closerange(but + 1, maxfds)
+            else:
+                for i in range(3, maxfds):
+                    if i == but:
+                        continue
+                    try:
+                        os.close(i)
+                    except:
+                        pass
 
 
 class TimedProc(object):
@@ -42,7 +78,7 @@ class TimedProc(object):
             raise salt.exceptions.TimedProcTimeoutError('Error: timeout {0} must be a number'.format(self.timeout))
 
         try:
-            self.process = subprocess.Popen(args, **kwargs)
+            self.process = FdPopen(args, **kwargs)
         except (AttributeError, TypeError):
             if not kwargs.get('shell', False):
                 if not isinstance(args, (list, tuple)):
@@ -72,7 +108,7 @@ class TimedProc(object):
                 # problems with subprocess.
                 kwargs['env'] = salt.utils.data.encode_dict(kwargs['env'])
             args = salt.utils.data.decode(args)
-            self.process = subprocess.Popen(args, **kwargs)
+            self.process = FdPopen(args, **kwargs)
         self.command = args
 
     def run(self):
